admin

+++ app/Http/Controllers/Admin/ContentController.php
@@ -1,23 +1,18 @@
 <?php
-
 namespace App\Http\Controllers\Admin;
-
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\File;
 use App\Models\Content;
 use App\Http\Services\Menu\MenuService;
 use App\Http\Requests\Content\StoreRequest;
-
 class ContentController extends Controller
 {
     public function index()
     {
         $title = 'Content';
-        $data = Content::search()->paginate(15);
-
+        $data = Content::orderBy('menu_id', 'ASC')->orderBy('priority')->search()->paginate(50);
         return view('admin.content.index', compact('data', 'title'));
     }
-
     public function create()
     {
         $title = 'Create Content';
@@ -51,7 +46,6 @@ class ContentController extends Controller
             return redirect()->route('content.index')->with('success', 'Add content success');
         }
     }
-
     public function update(StoreRequest $request, Content $content)
     {
         $old = $content->photo;
@@ -70,7 +64,6 @@ class ContentController extends Controller
             }
             $request->merge(['photo' => $fileName]);
         }
-
         if ($request->hasFile('mobile')) {
             if ($request->mobile->isValid()) {
                 $file = $request->mobile;
@@ -81,20 +74,23 @@ class ContentController extends Controller
             }
             $request->merge(['photo_mobile' => $fileName]);
         }
-
-        if ($request->photo && file_exists($path . '/' . $old)) {
+        if ($request->photo && $old && file_exists($path . '/' . $old)) {
             unlink($path . '/' . $old);
-        } elseif(empty($request->photo)) {
+        } elseif (empty($request->photo)) {
             $request->merge(['photo' => $old]);
         }
-
         if ($request->photo_mobile && $old_mobile && file_exists($path . '/' . $old_mobile)) {
             unlink($path . '/' . $old_mobile);
         } elseif (empty($request->photo_mobile)) {
             $request->merge(['photo_mobile' => $old_mobile]);
         }
-
         $content->update($request->only('title', 'photo', 'photo_mobile', 'description', 'menu_id', 'content_section', 'priority'));
         return redirect()->route('content.index')->with('success', 'Update content success');
+    }
+    public function destroy(Content $content)
+    {
+        if ($content->delete())
+            return redirect()->route('content.index')->with('success', 'Delete global network success');
+        return redirect()->route('content.index')->with('error', 'Errors');
     }
 }
